Rename pusher/pusher.py -> pusher/task.py; complete task.py convention
Establishes the OSS-wide convention that every workflow task package
uses task.py as its primary module (directory names the concept).
This follows the pattern set in PR #1426 for tabular_trainer and
mirrors the internal SDK convention universally.

- pusher.py -> task.py
- __init__.py import updated accordingly

Closes #1428
